SILVER-270 - Add support for GitHub Apps
The action now supports a client ID and client secret as authentication. This is usually what's needed for authenticating as a GitHub App.

I wanted to keep support for PAT tokens as I reckon in some places - both within Simply Business and in the open source community - people might still find it easier to use.